:sparkles: Create resource manager
The resource manager takes loaders, paired with regex keys. Resources
are then loaded simply by their file path, where the file extension is
matched with the regex keys, after which the resource is loaded with the
respective loader. Once a resource has been loaded, it can be fetched
with get() from the resource manager.

As an example, TextLoader has been implemented, which loads `.txt` files
as raw text.